refactor(heat_transfer): Step 1 - Move dataclasses to network.components
- Move SmoothModel, RibbedModel, DimpledModel, PinFinModel, ImpingementModel from heat_transfer.py to network/components.py
- Move ConvectiveSurface and WallConnection from heat_transfer.py to network/components.py
- Add backward-compatible re-exports in heat_transfer.py with __all__ to prevent ruff F401 errors
- Export new classes from network/__init__.py
- Add missing math import to components.py
- Keep compute_coupling method in WallConnection for backward compatibility
- Add ruff noqa comment for E402 (import not at top)
- All tests pass (18/18 total)
- Example script continues to work

This aligns with design document placement of dataclasses in components.py
while maintaining backward compatibility through re-exports.
